src/webapps/requirements.txt
Django==1.11.7
wheel==0.30.0
gunicorn==19.7.1
psycopg2==2.7.3.2
pillow
django-storages
boto
djangorestframework
django-webpack-loader
